\name{check.format}
\alias{check.format}
\title{
Check of TA (table A, hauls data), TB (table B, catches data), and TC (table C, biological data) tables' fields format}
\description{
The function checks whether TA (table A, hauls data), TB (table B, catches data), and TC (table C, biological data) tables have the currect fields. Missing columns or not expected fields could generate errors in the merging functions. \code{check.format} helps to identify the presence of unexpected or missing columns generating warnings.
}
\usage{
check.format(TA, TB, TC)
}
\arguments{
  \item{TA}{
data frame of hauls data (TA,  table A)}
  \item{TB}{
data frame of the catches data (TB,  table B)}
  \item{TC}{
data frame of the biological data (TC,  table C)}
}
\value{The function returns messages of the inconsistences detected in the names and in the order of the columns of the TA (table A, hauls data), TB (table B, catches data) and TC (table C, biological data) tables. In case no message is generated by the function the format of the data frames is assumed to be currect.}

\author{
Walter Zupa}

\examples{
check.format(TA, TB, TC)
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ format }% use one of  RShowDoc("KEYWORDS")
\keyword{ error }% __ONLY ONE__ keyword per line
